FlyMeBot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
#botbuilder-dialogs>=4.14.0
#botbuilder-ai>=4.14.0
datatypes-date-time>=1.0.0.a2

asyncio>=3.4.3
requests==2.23.0
botbuilder-core>=4.14.1
botframework-connector>=4.14.0
botbuilder-schema>=4.14.1
botbuilder-dialogs>=4.14.1
botbuilder-ai>=4.14.1
botbuilder-applicationinsights>=4.14.1
botbuilder-integration-applicationinsights-aiohttp>=4.14.1
azure-cognitiveservices-language-luis>=0.2.0
msrest>=0.6.10
aiohttp
python-dotenv